Removed old thesauri files
