+++ app/react/Settings/index.js
@@ -2,7 +2,6 @@
 import EntityTypesList from './components/EntityTypesList';
 import FiltersForm from './components/FiltersForm';
 import { RelationTypesList } from './components/RelationTypesList';
-import ThesauriList from './components/ThesauriList';
 import { Settings } from './Settings';
 import SettingsAPI from './SettingsAPI';
 import { Dashboard } from './components/Dashboard';
@@ -16,7 +15,6 @@ export {
   PreserveSettings,
   EntityTypesList,
   RelationTypesList,
-  ThesauriList,
   FiltersForm,
   Dashboard,
 };
